requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
Django==3.2.4
django-autoslug==1.9.8
django-cities-light==3.8.1
django-ckeditor==6.1.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-js-asset==1.2.2
django-otp==1.0.6
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
django-social-share==2.2.1
django-tag-parser==3.2
djangorestframework==3.12.4
idna==3.2
phonenumbers==8.12.25
Pillow==8.2.0
progressbar2==3.53.1
PyJWT==1.7.1
python-akismet==0.4.2
python-utils==2.5.6
pytz==2021.1
qrcode==7.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.62.0
Unidecode==1.2.0
urllib3==1.26.6
